requirements.txt
dbus-python==1.2.16
paho-mqtt==1.5.1
pycairo==1.20.0
PyGObject==3.38.0
